Moved getRouterForQuery to its own util
