+++ src/routes/views/details/awsDetails/awsDetails.tsx
@@ -21,9 +21,7 @@ import { NoProviders } from 'routes/state/noProviders';
 import { NotAvailable } from 'routes/state/notAvailable';
 import { ExportModal } from 'routes/views/components/export';
 import { getGroupByOrgValue, getGroupByTagKey } from 'routes/views/utils/groupBy';
-import { filterProviders, hasCurrentMonthData } from 'routes/views/utils/providers';
 import {
-  getRouteForQuery,
   handleCostTypeSelected,
   handleCurrencySelected,
   handleFilterAdded,
@@ -31,7 +29,9 @@ import {
   handlePerPageSelect,
   handleSetPage,
   handleSort,
-} from 'routes/views/utils/queryUpdate';
+} from 'routes/views/utils/handles';
+import { filterProviders, hasCurrentMonthData } from 'routes/views/utils/providers';
+import { getRouteForQuery } from 'routes/views/utils/query';
 import { createMapStateToProps, FetchStatus } from 'store/common';
 import { featureFlagsSelectors } from 'store/featureFlags';
 import { providersQuery, providersSelectors } from 'store/providers';
