+++ src/routes/views/details/components/breakdown/breakdownBase.tsx
@@ -14,8 +14,8 @@ import { Loading } from 'routes/state/loading';
 import { NoData } from 'routes/state/noData';
 import { NoProviders } from 'routes/state/noProviders';
 import { NotAvailable } from 'routes/state/notAvailable';
+import { handleCostTypeSelected, handleCurrencySelected } from 'routes/views/utils/handles';
 import { hasCurrentMonthData } from 'routes/views/utils/providers';
-import { handleCostTypeSelected, handleCurrencySelected } from 'routes/views/utils/queryUpdate';
 import { FetchStatus } from 'store/common';
 import type { reportActions } from 'store/reports';
 import type { CostTypes } from 'utils/costType';
