+++ src/routes/views/utils/handles.ts
@@ -1,21 +1,9 @@
-import type { Location } from '@remix-run/router';
 import type { Query } from 'api/queries/query';
-import { getQueryRoute } from 'api/queries/query';
 import type { RouteComponentProps } from 'utils/router';
 
 import type { Filter } from './filter';
 import { addFilterToQuery, removeFilterFromQuery } from './filter';
-
-export const getRouteForQuery = (query: Query, location: Location, reset: boolean = false) => {
-  // Reset pagination
-  if (reset) {
-    query.filter = {
-      ...query.filter,
-      offset: 0,
-    };
-  }
-  return `${location.pathname}?${getQueryRoute(query)}`;
-};
+import { getRouteForQuery } from './query';
 
 export const handleCurrencySelected = (
   query: Query,
